Rename the package to remove the underscore.
