+++ modules/avail_creator/src/main/java/org/hawkular/component/availcreator/MetricReceiver.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.component.avail_creator;
+package org.hawkular.component.availcreator;
 
 import com.google.gson.GsonBuilder;
